7/31/24 stuck with chunk mesh buider but hey. refactor include folder @@

+++ Source/Model.cpp
@@ -28,7 +28,7 @@ void Model::AddData(const Mesh& mesh) {
 		DeleteData();
 	}
 
-	m_indicesCount = mesh.indices.size();
+	m_indicesCount = (GLuint)mesh.indices.size();
 
 	// regenerate and bind data
 	m_vao.generateVao();
